queries and types have been moved to separate files
