+++ src/components/widgets/Like/Like.tsx
@@ -4,13 +4,12 @@ import { Button } from "@/components/shared/Button";
 
 import Heart from "@/assets/icons/heart.svg";
 
-import classes from "./Like.module.css";
+import { LikeProps } from "./types/LikeProps";
+
 import useLike from "./hooks/useLike";
 
-interface LikeProps {
-    event_id: string;
-    user_id: string;
-}
+import classes from "./Like.module.css";
+
 export const Like: FC<LikeProps> = ({ event_id, user_id }) => {
     const { handleLike, isLiked } = useLike({ event_id, user_id });
     return (
