+++ src/components/widgets/Like/hooks/useLike.ts
@@ -1,46 +1,12 @@
 import { useMutation, useQuery } from "@apollo/client";
-import gql from "graphql-tag";
 
-interface useLikeProps {
-    event_id: string;
-    user_id: string;
-}
+import { LIKED } from "@/app/api/graphql/queries/liked";
+import { LIKE_MUTATION } from "@/app/api/graphql/mutations/like";
 
-type ILike = {
-    _id: string;
-    event_id: string;
-    user_id: string;
-    isLiked: boolean;
-    createdAt?: string;
-    updatedAt?: string;
-};
+import { ILike } from "@/components/widgets/Like/types/Ilike";
+import { LikeProps } from "@/components/widgets/Like/types/LikeProps";
 
-const LIKE_MUTATION = gql`
-    mutation Mutation($eventId: ID!, $userId: ID!) {
-        like(event_id: $eventId, user_id: $userId) {
-            _id
-            event_id
-            user_id
-            isLiked
-            createdAt
-            updatedAt
-        }
-    }
-`;
-
-const LIKED = gql`
-    query Liked($eventId: ID, $userId: ID) {
-        liked(event_id: $eventId, user_id: $userId) {
-            _id
-            event_id
-            user_id
-            isLiked
-            createdAt
-            updatedAt
-        }
-    }
-`;
-const useLike = ({ event_id, user_id }: useLikeProps) => {
+const useLike = ({ event_id, user_id }: LikeProps) => {
     const [likeMutation] = useMutation(LIKE_MUTATION);
     const { data, refetch } = useQuery<{ liked: ILike | null }>(LIKED, {
         variables: { eventId: event_id, userId: user_id },
