+++ src/components/widgets/Like/types/Ilike.ts
@@ -1,0 +1,8 @@
+export type ILike = {
+    _id: string;
+    event_id: string;
+    user_id: string;
+    isLiked: boolean;
+    createdAt?: string;
+    updatedAt?: string;
+};
