+++ src/components/widgets/Like/types/LikeProps.ts
@@ -1,0 +1,4 @@
+export interface LikeProps {
+    event_id: string;
+    user_id: string;
+}
